# openapi.json
{"openapi":"3.0.3","info":{"title":"tracker-backend","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/activities":{"get":{"tags":["crate"],"operationId":"list_activities","responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivitySummary"}}}}}}},"post":{"tags":["crate"],"operationId":"create_activity","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateActivity"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivitySummary"}}}}}}},"/activities/{activity_id}/entries/{entry_id}":{"put":{"tags":["crate"],"operationId":"update_entry","parameters":[{"name":"activity_id","in":"path","description":"Activity ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"entry_id","in":"path","description":"Entry ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEntry"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogEntry"}}}},"404":{"description":"Not found"}}},"delete":{"tags":["crate"],"operationId":"delete_entry","parameters":[{"name":"activity_id","in":"path","description":"Activity ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"entry_id","in":"path","description":"Entry ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Deleted"},"404":{"description":"Not found"}}}},"/activities/{id}":{"get":{"tags":["crate"],"operationId":"get_activity","parameters":[{"name":"id","in":"path","description":"Activity ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityDetail"}}}},"404":{"description":"Not found"}}},"delete":{"tags":["crate"],"operationId":"delete_activity","parameters":[{"name":"id","in":"path","description":"Activity ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Deleted"},"404":{"description":"Not found"}}},"patch":{"tags":["crate"],"operationId":"rename_activity","parameters":[{"name":"id","in":"path","description":"Activity ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateActivity"}}},"required":true},"responses":{"204":{"description":"Renamed"},"404":{"description":"Not found"}}}},"/activities/{id}/entries":{"post":{"tags":["crate"],"operationId":"create_entry","parameters":[{"name":"id","in":"path","description":"Activity ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEntry"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogEntry"}}}}}}},"/activities/{id}/stats":{"get":{"tags":["crate"],"operationId":"get_stats","parameters":[{"name":"id","in":"path","description":"Activity ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"period","in":"query","description":"day | week | month","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityStats"}}}}}}}},"components":{"schemas":{"ActivityDetail":{"type":"object","required":["id","name","created_at","last_entries","field_schema"],"properties":{"created_at":{"type":"integer","format":"int64"},"field_schema":{"type":"array","items":{"$ref":"#/components/schemas/FieldSchema"},"description":"Inferred from last entry; empty if no entries exist yet"},"id":{"type":"integer","format":"int64"},"last_entries":{"type":"array","items":{"$ref":"#/components/schemas/LogEntry"},"description":"Last 10 entries, most recent first"},"name":{"type":"string"}}},"ActivityStats":{"type":"object","required":["period","counts","field_stats"],"properties":{"counts":{"type":"array","items":{"$ref":"#/components/schemas/DailyCount"}},"field_stats":{"type":"array","items":{"$ref":"#/components/schemas/FieldStat"}},"period":{"type":"string"}}},"ActivitySummary":{"type":"object","required":["id","name","created_at","entry_count"],"properties":{"created_at":{"type":"integer","format":"int64"},"entry_count":{"type":"integer","format":"int64"},"id":{"type":"integer","format":"int64"},"last_logged_at":{"type":"integer","format":"int64","nullable":true},"name":{"type":"string"}}},"CreateActivity":{"type":"object","required":["name"],"properties":{"name":{"type":"string"}}},"CreateEntry":{"type":"object","required":["values"],"properties":{"comment":{"type":"string","nullable":true},"logged_at":{"type":"integer","format":"int64","description":"Unix timestamp seconds; omit to default to now","nullable":true},"values":{"type":"array","items":{"$ref":"#/components/schemas/FieldValue"}}}},"DailyCount":{"type":"object","required":["date","count"],"properties":{"count":{"type":"integer","format":"int64"},"date":{"type":"string"}}},"FieldSchema":{"type":"object","required":["label","last_value"],"properties":{"label":{"type":"string"},"last_value":{"type":"number","format":"double"}}},"FieldStat":{"type":"object","required":["label","data"],"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/FieldStatPoint"}},"label":{"type":"string"}}},"FieldStatPoint":{"type":"object","required":["date"],"properties":{"avg":{"type":"number","format":"double","nullable":true},"date":{"type":"string"},"sum":{"type":"number","format":"double","nullable":true}}},"FieldValue":{"type":"object","required":["label","value"],"properties":{"label":{"type":"string"},"value":{"type":"number","format":"double"}}},"LogEntry":{"type":"object","required":["id","activity_id","logged_at","values"],"properties":{"activity_id":{"type":"integer","format":"int64"},"comment":{"type":"string","nullable":true},"id":{"type":"integer","format":"int64"},"logged_at":{"type":"integer","format":"int64"},"values":{"type":"array","items":{"$ref":"#/components/schemas/FieldValue"}}}},"Period":{"type":"string","enum":["day","week","month"]},"UpdateEntry":{"type":"object","required":["values","logged_at"],"properties":{"activity_name":{"type":"string","description":"If provided, renames the parent activity (affects all its entries)","nullable":true},"comment":{"type":"string","nullable":true},"logged_at":{"type":"integer","format":"int64"},"values":{"type":"array","items":{"$ref":"#/components/schemas/FieldValue"}}}}}}}